+++ jot/Commands/Schemas/ImportMaps/ImportMapRenameCommand.cs
@@ -1,0 +1,71 @@
+/*
+Figment
+Copyright (C) 2025  Sean McElroy
+
+This program is free software: you can redistribute it and/or modify
+it under the terms of the GNU Affero General Public License as published by
+the Free Software Foundation, either version 3 of the License, or
+(at your option) any later version.
+
+This program is distributed in the hope that it will be useful,
+but WITHOUT ANY WARRANTY; without even the implied warranty of
+MERCHANTABILITY or FITNESS FOR A PARTICULAR PURPOSE.  See the
+GNU Affero General Public License for more details.
+
+You should have received a copy of the GNU Affero General Public License
+along with this program.  If not, see <http://www.gnu.org/licenses/>.
+*/
+
+using Figment.Common.Errors;
+using Spectre.Console.Cli;
+
+namespace jot.Commands.Schemas.ImportMaps;
+
+/// <summary>
+/// Changes the name of a schema.
+/// </summary>
+public class ImportMapRenameCommand : SchemaCancellableAsyncCommand<ImportMapRenameCommandSettings>
+{
+    /// <inheritdoc/>
+    public override async Task<int> ExecuteAsync(CommandContext context, ImportMapRenameCommandSettings settings, CancellationToken cancellationToken)
+    {
+        var (tgs, schema, ssp) = await TryGetSchema(settings, cancellationToken);
+        if (tgs != Globals.GLOBAL_ERROR_CODES.SUCCESS)
+        {
+            return (int)tgs;
+        }
+
+        if (string.IsNullOrWhiteSpace(settings.NewName))
+        {
+            AmbientErrorContext.Provider.LogError("Name of an import map cannot be empty.");
+            return (int)Globals.GLOBAL_ERROR_CODES.ARGUMENT_ERROR;
+        }
+
+        var importMap = schema!.ImportMaps.FirstOrDefault(i => string.Compare(i.Name, settings.ImportMapName, StringComparison.InvariantCultureIgnoreCase) == 0);
+
+        if (importMap == null)
+        {
+            AmbientErrorContext.Provider.LogError($"Schema '{schema.Name}' does not have an import map named '{settings.ImportMapName}'.");
+            return (int)Globals.GLOBAL_ERROR_CODES.NOT_FOUND;
+        }
+
+        var oldName = importMap!.Name;
+        importMap.Name = settings.NewName.Trim();
+        var saved = await schema.SaveAsync(cancellationToken);
+        if (!saved)
+        {
+            if (settings.Verbose ?? false)
+            {
+                AmbientErrorContext.Provider.LogError($"Unable to save import map '{importMap.Name}' on schema '{schema.Name}' ({schema.Guid}).");
+            }
+            else
+            {
+                AmbientErrorContext.Provider.LogError($"Unable to save import map '{importMap.Name}' on schema '{schema.Name}'.");
+            }
+
+            return (int)Globals.GLOBAL_ERROR_CODES.SCHEMA_SAVE_ERROR;
+        }
+
+        return (int)Globals.GLOBAL_ERROR_CODES.SUCCESS;
+    }
+}