+++ jot/Commands/Schemas/ImportMaps/NewImportMapCommand.cs
@@ -1,3 +1,5 @@
+using System.Globalization;
+using CsvHelper;
 using Figment.Common;
 using Figment.Common.Errors;
 using Spectre.Console.Cli;
@@ -24,6 +26,13 @@ public class NewImportMapCommand : SchemaCancellableAsyncCommand<NewImportMapCom
             return (int)Globals.GLOBAL_ERROR_CODES.ARGUMENT_ERROR;
         }
 
+        if (string.IsNullOrWhiteSpace(settings.SampleFilePath)
+            || !File.Exists(settings.SampleFilePath))
+        {
+            AmbientErrorContext.Provider.LogError($"File path '{settings.SampleFilePath}' was not found.");
+            return (int)Globals.GLOBAL_ERROR_CODES.ARGUMENT_ERROR;
+        }
+
         var (tgs, schema, _) = await TryGetSchema(settings, cancellationToken);
         if (tgs != Globals.GLOBAL_ERROR_CODES.SUCCESS)
         {
@@ -37,6 +46,15 @@ public class NewImportMapCommand : SchemaCancellableAsyncCommand<NewImportMapCom
         }
 
         var importMap = new SchemaImportMap(settings.ImportMapName, settings.FileType);
+
+        switch (settings.FileType.ToLowerInvariant())
+        {
+            case "csv":
+                var csvFields = InferImportMapFieldsFromCsv(settings.SampleFilePath)
+                    .ToBlockingEnumerable(cancellationToken);
+                importMap.FieldConfiguration.AddRange(csvFields);
+                break;
+        }
 
         schema!.ImportMaps.Add(importMap);
         var saved = await schema.SaveAsync(cancellationToken);
@@ -56,5 +74,49 @@ public class NewImportMapCommand : SchemaCancellableAsyncCommand<NewImportMapCom
 
         AmbientErrorContext.Provider.LogDone($"Added new import map for '{importMap.Name}' to '{schema.Name}'.");
         return (int)Globals.GLOBAL_ERROR_CODES.SUCCESS;
+    }
+
+    /// <summary>
+    /// Infers the <see cref="SchemaImportField"/> objects from reading the header of a CSV file.
+    /// </summary>
+    /// <param name="filePath">Comma-separated value file with headers to read.</param>
+    /// <returns>An asynchronous enumerator that returns each read file field with no property mapped.</returns>
+    public static async IAsyncEnumerable<SchemaImportField> InferImportMapFieldsFromCsv(string filePath)
+    {
+        ArgumentException.ThrowIfNullOrWhiteSpace(filePath);
+
+        if (!File.Exists(filePath))
+        {
+            yield break;
+        }
+
+        using var reader = new StreamReader(filePath);
+        using var csv = new CsvReader(reader, CultureInfo.InvariantCulture);
+
+        if (!await csv.ReadAsync())
+        {
+            AmbientErrorContext.Provider.LogError($"Unable to read CSV file {filePath}");
+            yield break;
+        }
+
+        if (!csv.ReadHeader())
+        {
+            AmbientErrorContext.Provider.LogError($"Unable to read CSV file headers from {filePath}");
+            yield break;
+        }
+
+        if (csv.HeaderRecord == null)
+        {
+            yield break;
+        }
+
+        foreach (var header in csv.HeaderRecord)
+        {
+            yield return new SchemaImportField(null, header)
+            {
+                SkipRecordIfInvalid = true,
+                SkipRecordIfMissing = true,
+            };
+        }
     }
 }